Use MUnit in ResultSuite
